refactor: consolidate parser code
