+++ src/server/features/completion/completion.ts
@@ -5,16 +5,16 @@ import {
 } from "vscode-languageserver";
 import type { CompletionItem } from "vscode-languageserver";
 import type { TextDocument } from "vscode-languageserver-textdocument";
-import { sassBuiltInModules } from "../../sass-built-in-modules";
-import type { SassBuiltInModule } from "../../sass-built-in-modules";
-import type { ISettings } from "../../settings";
-import type StorageService from "../../storage";
 import type {
 	IScssDocument,
 	ScssForward,
 	ScssImport,
 	ScssUse,
-} from "../../symbols";
+} from "../../parser";
+import { sassBuiltInModules } from "../../sass-built-in-modules";
+import type { SassBuiltInModule } from "../../sass-built-in-modules";
+import type { ISettings } from "../../settings";
+import type StorageService from "../../storage";
 import { asDollarlessVariable } from "../../utils/string";
 import { createCompletionContext } from "./completion-context";
 import type { CompletionContext } from "./completion-context";
