+++ src/test/features/completion.spec.ts
@@ -2,9 +2,9 @@ import assert from "assert";
 import { CompletionItemKind, SymbolKind } from "vscode-languageserver";
 import type { CompletionList } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
-import { ScssDocument } from "../../server/document";
 import { doCompletion } from "../../server/features/completion";
 import { parseStringLiteralChoices } from "../../server/features/completion/completion-utils";
+import { ScssDocument } from "../../server/parser";
 import { sassBuiltInModules } from "../../server/sass-built-in-modules";
 import type { ISettings } from "../../server/settings";
 import StorageService from "../../server/storage";
