+++ src/test/features/go-definition.spec.ts
@@ -1,8 +1,8 @@
 import assert from "assert";
 import { SymbolKind } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
-import { ScssDocument } from "../../server/document";
 import { goDefinition } from "../../server/features/go-definition";
+import { ScssDocument } from "../../server/parser";
 import StorageService from "../../server/storage";
 import * as helpers from "../helpers";
