+++ src/test/features/hover.spec.ts
@@ -2,8 +2,8 @@ import assert from "assert";
 import { MarkupKind, SymbolKind } from "vscode-languageserver";
 import type { Hover } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
-import { ScssDocument } from "../../server/document";
 import { doHover } from "../../server/features/hover";
+import { ScssDocument } from "../../server/parser";
 import StorageService from "../../server/storage";
 import * as helpers from "../helpers";
