+++ src/test/features/workspace-symbol.spec.ts
@@ -1,8 +1,8 @@
 import assert from "assert";
 import { SymbolKind } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
-import { ScssDocument } from "../../server/document";
 import { searchWorkspaceSymbol } from "../../server/features/workspace-symbol";
+import { ScssDocument } from "../../server/parser";
 import StorageService from "../../server/storage";
 
 const storage = new StorageService();
