+++ src/test/scanner.spec.ts
@@ -5,10 +5,10 @@ import { Stats } from "@nodelib/fs.macchiato";
 import { isMatch } from "micromatch";
 import { stub, SinonStub } from "sinon";
 import { URI } from "vscode-uri";
-import ScannerService from "../../server/scanner";
-import StorageService from "../../server/storage";
-import * as fsUtils from "../../server/utils/fs";
-import * as helpers from "../helpers";
+import ScannerService from "../server/scanner";
+import StorageService from "../server/storage";
+import * as fsUtils from "../server/utils/fs";
+import * as helpers from "./helpers";
 
 describe("Services/Scanner", () => {
 	describe(".scan", () => {
